requirements.txt
ttkbootstrap>=1.10.1
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.5.0
torchvision>=0.20.0
